Various bugfixes to get JSON working
The biggest brainfuck here was that Support.checkNotNull was wrong. It
threw a NullPointerException if the object was != null. Fun.

Other stuff include fixing the datamodel to avoid parsing errors due to
input error on the eventsystem. Some values can be null, and ints and
booleans do not like that at all.