Add basic GUI to application
Look, I don't want to discuss what's going on in this, well, horrible
code. Because of sorting and bucketing issues with events I had to
implement my own FestivalDay-class with custom compareTo and hashcode.
Then, I had to bucketize using TreeMap and finally construct a hell of a
lot of custom classes and views to render it properly.

However, shit works.